Centralize package categorization logic
