+++ tools/smithy-rs-tool-common/src/lib.rs
@@ -5,4 +5,5 @@
 
 pub mod git;
 pub mod github_actions;
+pub mod package;
 pub mod shell;
